argument 'times' removed from solver function
The argument isn't really needed - the same result can be achieved
by invoking `set_times()` on the scenario. In addition, it made
the inner workings of `solver()` more complicated than necessary.
However, checking the consistency of the output times vector is
now the responsibility of `set_times()` and, consequently,
`set_exposure()` has become more picky on what it deems valid
input. An acceptable loss of comfort.