GUTS-RED-IT: remove dependence of output time points on survival
Because surival `S` was calculated for the IT model after completed
simulation based on returned Damage (`D`) values, lethality could be
underestimated if output times did not include the periods of
(highest) exposure. To avoid this side-effect, the cumulative
maximum of `D` is now calculated as part of the ODE of the IT model.

However, the cumulative maximum cannot be approximated with
unlimited precision with the means provided by deSolve's compiled
code ODEs. Nonethless, deviations from the exact result will be so
small that they have no impact on calculated effects and EPx values
and will very likely be indistinguishable from numerical errors.